fix prebuildConfig.sh script when no config

+++ app/bin/prebuildConfig.sh
@@ -50,7 +50,9 @@ if test -f icons/icon-48.webp; then
   mv icons ./public/assets/icons
 fi
 
-## copy the icons over to web/ as well
-rm -rf ../web/public/assets/icons
-mkdir -p ../web/public/assets/
-cp -r ./public/assets/icons ../web/public/assets/icons
+## copy the icons over to web/ as well (check they exist first)
+if test -f ./public/assets/icons/icon-48.webp; then
+  rm -rf ../web/public/assets/icons
+  mkdir -p ../web/public/assets/
+  cp -r ./public/assets/icons ../web/public/assets/icons
+fi
